requirements.txt
python-docx
numpy
sentence-transformers
torch
huggingface_hub
pytest
pyqt5
mammoth
